Refactor: Restructure package layout
