+++ yagu-route/src/main/java/org/teamtuna/yaguroute/YaguRouteApplication.java
@@ -5,6 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class YaguRouteApplication {
+
     public static void main(String[] args) {
         SpringApplication.run(YaguRouteApplication.class, args);
     }
